pyproject.toml
[tool.poetry]
name = "workshop-api"
version = "0.1.0"
description = ""
authors = ["Fabio Cantarim <fabio.cantarim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
